refactor(ui): lift useProviderDefaults into its own file (todo #288)
use-composer-tuning.ts has to absorb the mid-session tuning-warning wiring and
sits close to the 300-line limit. The hook moves verbatim and is re-exported
from its old home, so every existing import path keeps working.
